nd0044/capstone/starter/requirements.txt
aniso8601==6.0.0
astroid==2.2.5
Click==8.0.1
cryptography==37.0.4
ecdsa==0.13.2
flask==1.1.2
Flask-Cors==3.0.10
flask_migrate==3.1.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
flask_sqlalchemy==2.5.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.4.0
MarkupSafe==2.0.1
mock==4.0.3
psycopg2==2.9.3
pycryptodome==3.15.0
PyJWT==2.4.0
pylint==2.3.1
python-jose==3.3.0
pytz==2019.1
six==1.16.0
SQLAlchemy==1.4.18
typed-ast==1.4.2
Werkzeug==2.0.1
wrapt==1.11.1